Allow for a shared authentication header
Previously, the authentication would always fail if the prefix on the
HTTP header did not match that expected by this package. This made it
impossible to use the same header, for example `Authorization`, for more
than one type of authentication. This allows for the developer to
configure if this strict check should be enforced.

If the header is defined as a shared header, then receiving a different
prefix than that defined for personal access tokens will not cause an
error, assuming that a different middleware or authenticator will handle
it instead.

The default behavior is the strict check that will throw an error if
a different prefix value is provided.