Clean up test flow to reduce duplication
Realized that the django version matrix in the GitHub action wasn't
being used when running the nox sessions. So I was duplicating tests
against different versions of Django in the same versions of python.
THis simplifies the action to just have one build matrix for each
version of Python and OS, and within that build, it will run against
each version of Django using nox.